refactor(registry): split canonicalize.rs into module
Split canonicalize.rs (1155 LOC) into:
- mod.rs: façade, orchestration, re-exports, tests
- errors.rs: CanonicalizeError, CanonicalizeResult, MAX_* constants
- yaml.rs: parse_yaml_strict, pre_scan_yaml, yaml_to_json
- json.rs: to_canonical_jcs_bytes (JCS/RFC 8785)
- digest.rs: sha256_prefixed

Public API unchanged. All tests pass.
